Adding file detailing the new mapping of HyCal
This spreadsheet contains the complete mapping of the modules in HyCal to their VPCB connector slots, those slots to the pins atop HyCal, those pins to the cable numbers, and the cables to the fADCs. The first sheet has all of this information, with the following sheets having the denoted individual maps included updated images of the VPCB connections.